==> 0_Overview_of_samples.Rmd <==
---
title: "Inital overview of samples"
output:
  github_document:
    toc: true
editor_options:
  chunk_output_type: console
---

## Load packages
```{r message=FALSE, warning=FALSE, error=TRUE}
library(dplyr)
library(tidyverse)
library(ggplot2)
library(ggpubr)

path <- "C:/Users/FX76TZ/OneDrive - Aalborg Universitet/Mikrobiom korrelation/"

packageVersion("dplyr")
packageVersion("tidyverse")
packageVersion("ggplot2")
packageVersion("ggpubr")
```


## Preperation of the American Gut Project (PRJEB11419)
```{r}
df <- read.csv(paste0(path, "Metadata/test.txt"))
table(df$sample_type)
df <- df %>% filter(sample_type %in% c("stool", "Stool", "feces"))

table(df$ibd_diagnosis_refined)

#Remove individuals with other diseases 
hc <- df %>% filter(ibd == "I do not have this condition") %>% filter(liver_disease == "I do not have this condition") %>% filter(ibs == "I do not have this condition") %>% filter(migraine == "I do not have this condition") %>% filter(appendix_removed %in% c("false", "No")) %>% filter(asd == "I do not have this condition") %>% filter(autoimmune == "I do not have this condition")%>% filter(tonsils_removed %in% c("false", "No"))%>% filter(diabetes == "I do not have this condition") %>% filter(antibiotic_history == "I have not taken antibiotics in the past year.") %>% filter(!(cancer_treatment %in% c("Chemotherapy", "Radiation therapy", "Surgery only"))) %>% filter(skin_condition == "I do not have this condition") %>% filter(thyroid == "I do not have this condition")%>% filter(add_adhd == "I do not have this condition")%>% filter(asd == "I do not have this condition") %>% filter(fungal_overgrowth == "I do not have this condition") %>% filter(gluten != "I was diagnosed with celiac disease") %>% filter(gluten != "I was diagnosed with gluten allergy (anti-gluten IgG), but not celiac disease")%>% filter(clinical_condition == "I do not have this condition") %>% filter(!(pregnant %in% c("Yes", "true")))

cd <- df %>% filter(ibd_diagnosis_refined %in% c("Colonic Crohn's Disease", "Ileal and Colonic Crohn's Disease", "Ileal Crohn's Disease"))
uc <- df %>% filter(ibd_diagnosis_refined == "Ulcerative colitis")

#Remove duplicated individuals
cd <- cd[!duplicated(cd$Submitter_Id), ]
uc <- uc[!duplicated(uc$Submitter_Id), ]
hc <- hc[!duplicated(hc$Submitter_Id), ]

uc$Disease <- "UC"
cd$Disease <- "CD"
hc$Disease <- "HC"
uc$region <- "V4"
uc$chem <- "NA"
uc$sex <- "NA"
uc$BMI <- as.numeric(uc$weight_kg)/(as.numeric(uc$height_cm)/100)^2
cd$region <- "V4"
cd$chem <- "NA"
cd$sex <- "NA"
cd$BMI <- as.numeric(cd$weight_kg)/(as.numeric(cd$height_cm)/100)^2
hc$region <- "V4"
hc$chem <- "NA"
hc$sex <- "NA"
hc$BMI <- as.numeric(hc$weight_kg)/(as.numeric(hc$height_cm)/100)^2


uc <- uc %>% select(Run, AvgSpotLen, Bases,BioProject, BioSample, Bytes, Center.Name, ENA.LAST.UPDATE..run., Experiment, Disease, geo_loc_name_country, geo_loc_name_country_continent, geo_loc_name, Instrument, Library.Name, LibraryLayout, Platform, Sample.Name, SRA.Study, region, chem, Age_years, BMI, sex, smoking_frequency)
cd <- cd %>% select(Run, AvgSpotLen, Bases,BioProject, BioSample, Bytes, Center.Name, ENA.LAST.UPDATE..run., Experiment, Disease, geo_loc_name_country, geo_loc_name_country_continent, geo_loc_name, Instrument, Library.Name, LibraryLayout, Platform, Sample.Name, SRA.Study, region, chem, Age_years, BMI, sex, smoking_frequency)
hc <- hc %>% select(Run, AvgSpotLen, Bases,BioProject, BioSample, Bytes, Center.Name, ENA.LAST.UPDATE..run., Experiment, Disease, geo_loc_name_country, geo_loc_name_country_continent, geo_loc_name, Instrument, Library.Name, LibraryLayout, Platform, Sample.Name, SRA.Study, region, chem, Age_years, BMI, sex, smoking_frequency)

df <- rbind(uc, cd, hc)
write.csv(df, file = paste0(path, "/Metadata/american_df.txt"), quote = FALSE, row.names = FALSE)
```

## Overview of IBD studies with 16S-seq
```{r}
df <- read.csv(paste0(path, "Metadata/Metadata_IBD_project.csv"), sep = ";")
american <- read.csv(file = paste0(path, "/Metadata/american_df_updated.txt"), sep="\t")
american <- american %>% dplyr::select(-chem)
colnames(american) <- colnames(df)
df_total <- rbind(df, american)
df_total[df_total==""]<-NA

df_total <- df_total %>% transform(AvgSpotLen = as.numeric(AvgSpotLen), Bases = as.numeric(Bases), Bytes = as.numeric(Bytes), Host_Age = as.numeric(Host_Age), host_body_mass_index = as.numeric(host_body_mass_index))
df <- df_total

table(df$Disease, useNA = "always")
length(unique(df$SRA.Study))
table(df$host_sex, useNA = "always")
table(df$smoking_status, useNA = "always")
unique(df$geo_loc_name_country)
unique(df$geo_loc_name_country_continent)

#Split by disease -> check samples, gender, age, BMI, smoking
df_UC <- df %>% filter(Disease == "UC")
df_CD <- df %>% filter(Disease == "CD")
df_HC <- df %>% filter(Disease == "HC")
df_NA <- df %>% filter(is.na(Disease))

table(df_UC$host_sex, useNA = "always")
table(df_CD$host_sex, useNA = "always")
table(df_HC$host_sex, useNA = "always")
table(df_NA$host_sex, useNA = "always")

summary(df_UC$Host_Age)
summary(df_CD$Host_Age)
summary(df_HC$Host_Age)
summary(df_NA$Host_Age)

summary(df_UC$host_body_mass_index)
summary(df_CD$host_body_mass_index)
summary(df_HC$host_body_mass_index)
summary(df_NA$host_body_mass_index)

table(df_UC$smoking_status, useNA = "always")
table(df_CD$smoking_status, useNA = "always")
table(df_HC$smoking_status, useNA = "always")
table(df_NA$smoking_status, useNA = "always")

df$smoking_status <- ifelse(df$smoking_status %in% c("active", "Daily", "Rarely (a few times/month)", "Regularly (3-5 times/week)", "smoker"), "smoker", ifelse(df$smoking_status %in% c("Never", "non-smoker"), "non-smoker", ifelse(df$smoking_status == "former", "former", NA)))
df$host_sex <- gsub("missing", NA, df$host_sex)
df$geo_loc_name_country <- gsub("United Kingdom", "UK", df$geo_loc_name_country)
df$geo_loc_name_country <- gsub("uncalculated", NA, df$geo_loc_name_country)
df$geo_loc_name_country <- gsub("The Netherlands", "Netherlands", df$geo_loc_name_country)
df$geo_loc_name_country_continent <- gsub("Australia", "Oceania", df$geo_loc_name_country_continent)
df$geo_loc_name_country_continent <- gsub("uncalculated", NA, df$geo_loc_name_country_continent)

write.table(df, file = paste0(path, "Metadata/Metadata_IBD_full.txt"), quote=F, row.names = F, sep="\t")

#Split by single/paired:
df_single <- df %>% filter(LibraryLayout == "SINGLE")
df_paired <- df %>% filter(LibraryLayout == "PAIRED")

#Printing commands for downloading samples from NCBI
single_projects <- unique(df_single$BioProject)
#To print names of all samples:
# for (j in single_projects){
# t <- ""
# df_sub <- df_single %>% filter(BioProject==j)
# for (i in df_sub$ï..Run){
# t <- paste0(t, " ", i)
# }
# print(j)
# print(t)
# }
# 
# paired_projects <- unique(df_paired$BioProject)
# for (j in paired_projects){
# t <- ""
# df_sub <- df_paired %>% filter(BioProject==j)
# for (i in df_sub$ï..Run){
# t <- paste0(t, " ", i)
# }
# print(j)
# print(t)
# }
```

### V4 region
```{r}
#V4
df_v4 <- df %>% filter(X16s_region =="V4")
table(df_v4$Disease, useNA = "always")
unique(df_v4$geo_loc_name_country)
unique(df_v4$geo_loc_name_country_continent)
length(unique(df_v4$BioProject))
```

### V3-V4 region
```{r}
#V3-V4
df_v34 <- df %>% filter(X16s_region =="V3-V4")
table(df_v34$Disease, useNA = "always")
unique(df_v34$geo_loc_name_country)
unique(df_v34$geo_loc_name_country_continent)
length(unique(df_v34$BioProject))
```

### V1-V3 region
```{r}
#V1-V3
df_v13 <- df %>% filter(X16s_region =="V1-V3")
table(df_v13$Disease, useNA = "always")
unique(df_v13$geo_loc_name_country)
unique(df_v13$geo_loc_name_country_continent)
length(unique(df_v13$BioProject))
```

### V1-V2 region
```{r}
#V1-V2
df_v12 <- df %>% filter(X16s_region =="V1-V2")
table(df_v12$Disease, useNA = "always")
unique(df_v12$geo_loc_name_country)
unique(df_v12$geo_loc_name_country_continent)
length(unique(df_v12$BioProject))
```

### V4-V5 region
```{r}
#V4-V5
df_v45 <- df %>% filter(X16s_region =="V4-V5")
table(df_v45$Disease, useNA = "always")
unique(df_v45$geo_loc_name_country)
unique(df_v45$geo_loc_name_country_continent)
length(unique(df_v45$BioProject))
```

### V3-V5 region
```{r}
#V3-V5
df_v35 <- df %>% filter(X16s_region =="V3-V5")
table(df_v35$Disease, useNA = "always")
unique(df_v35$geo_loc_name_country)
unique(df_v35$geo_loc_name_country_continent)
length(unique(df_v35$BioProject))
```

### V3 region
```{r}
#V3
df_v3 <- df %>% filter(X16s_region =="V3")
table(df_v3$Disease, useNA = "always")
unique(df_v3$geo_loc_name_country)
unique(df_v3$geo_loc_name_country_continent)
length(unique(df_v3$BioProject))
```

### V3-V6 region
```{r}
#V3-V6
df_v36 <- df %>% filter(X16s_region =="V3-V6")
table(df_v36$Disease, useNA = "always")
unique(df_v36$geo_loc_name_country)
unique(df_v36$geo_loc_name_country_continent)
length(unique(df_v36$BioProject))
```

### V3-V4, V6-V9 region
```{r}
#V2, V3, V4, V6-V7, V8, V9
df_vall <- df %>% filter(X16s_region =="V2, V3, V4, V6-V7, V8, V9")
table(df_vall$Disease, useNA = "always")
unique(df_vall$geo_loc_name_country)
unique(df_vall$geo_loc_name_country_continent)
length(unique(df_vall$BioProject))
```


## Plot of available data
```{r}
df_projects <- df %>% dplyr::group_by(X16s_region, Instrument) %>% summarise(samples = n(), projects = length(unique(BioProject)))
df_projects$X16s_region
df_projects[6, "X16s_region"] <- "All"

p1 <- ggplot(data=df_projects)+
  geom_bar(aes(x=reorder(X16s_region,-projects, sum), y=projects, fill=Instrument), stat="identity")+xlab("16S V-region")+ylab("Studies")+theme_classic()+ theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggplot(data=df_projects)+
  geom_bar(aes(x=reorder(X16s_region,-samples), y=samples), stat="identity")+xlab("16S V-region")+ylab("Total samples")+theme_classic()

df_projects <- df %>% dplyr::group_by(X16s_region, BioProject) %>% summarise(samples = n())
df_projects$X16s_region
df_projects[10, "X16s_region"] <- "All"

ggplot(data=df_projects)+
  geom_bar(aes(x=reorder(X16s_region,-samples, sum), y=samples, fill=BioProject), stat="identity")+xlab("16S V-region")+ylab("Total samples")+theme_classic()+theme(legend.position = "none")

df_projects <- df %>% dplyr::group_by(X16s_region, Disease) %>% summarise(samples = n())
df_projects$X16s_region
df_projects[8:9, "X16s_region"] <- "All"

p3 <- ggplot(data=df_projects)+
  geom_bar(aes(x=reorder(X16s_region,-samples, sum), y=samples, fill=Disease), stat="identity")+xlab("16S V-region")+ylab("Total samples")+theme_classic()+ theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

df_projects <- df %>% dplyr::group_by(X16s_region, Instrument) %>% summarise(samples = n())
df_projects[6, "X16s_region"] <- "All"

ggplot(data=df_projects)+
  geom_bar(aes(x=reorder(X16s_region,-samples, sum), y=samples, fill=Instrument), stat="identity")+xlab("16S V-region")+ylab("Total samples")+theme_classic()

df_disease <- df %>% dplyr::group_by(Disease, host_sex) %>% summarise(samples = n())

p4 <- ggplot(data=df_disease)+
  geom_bar(aes(x=reorder(Disease,-samples, sum), y=samples, fill=host_sex), stat="identity")+xlab("Disease status")+ylab("Total samples")+theme_classic()+labs(fill="Gender")

df_instrument <-  df %>% dplyr::group_by(Instrument, LibraryLayout) %>% summarise(samples = n())
p2 <- ggplot(data=df_instrument)+
  geom_bar(aes(x=reorder(Instrument,-samples, sum), y=samples, fill=LibraryLayout), stat="identity")+xlab("Instrument")+ylab("Total samples")+theme_classic()+ theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))+labs(fill="Layout")

ggarrange(p1, p2, p3, p4, labels=c("A", "B", "C", "D"))

```
